requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
bcrypt==3.1.7             # via paramiko
cffi==1.14.1              # via bcrypt, cryptography, pynacl
coverage==5.2.1           # via -r requirements.in
cryptography==3.0         # via paramiko
django-js-routes==0.1.3   # via -r requirements.in
django-rest-framework==0.1.0  # via -r requirements.in
django-seed==0.2.2        # via -r requirements.in
django-test-plus==1.4.0   # via -r requirements.in
django==3.1               # via -r requirements.in, django-js-routes, django-seed, djangorestframework, inertia-django
djangorestframework==3.11.1  # via django-rest-framework
fabric==2.5.0             # via -r requirements.in, patchwork
faker==4.1.1              # via django-seed
inertia-django==0.2.3     # via -r requirements.in
invoke==1.4.1             # via fabric
paramiko==2.7.1           # via fabric
patchwork==1.0.1          # via -r requirements.in
pycparser==2.20           # via cffi
pynacl==1.4.0             # via paramiko
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django
six==1.15.0               # via bcrypt, cryptography, pynacl, python-dateutil
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
